requirements.txt
tensorflow==2.16.1
tf_keras==2.16.0
keras==3.1.1
scikit-learn==1.4.0
pandas==2.2.0
PyYAML==6.0.1
peft==0.14.0
trl==0.8.6
transformers==4.47.1
xgboost==2.0.3
